chore: remove dead UI deps and unify USSD CSV parsing
Drop unused shadcn/loading components and consolidate CSV tokenizer/mapping into a shared ussdCsv module.
