--- web/lib/parseCsv.ts
@@ -1,76 +1,0 @@
-/**
- * Parse CSV text (RFC 4180) — handles quoted fields, escaped quotes, and commas.
- */
-export function parseCsv(text: string): string[][] {
-  const input = text.replace(/^\uFEFF/, '');
-  const rows: string[][] = [];
-  let row: string[] = [];
-  let field = '';
-  let inQuotes = false;
-  let i = 0;
-
-  const pushField = () => {
-    row.push(field.trim());
-    field = '';
-  };
-
-  const pushRow = () => {
-    pushField();
-    if (row.some((cell) => cell !== '')) {
-      rows.push(row);
-    }
-    row = [];
-  };
-
-  while (i < input.length) {
-    const char = input[i];
-
-    if (inQuotes) {
-      if (char === '"') {
-        if (input[i + 1] === '"') {
-          field += '"';
-          i += 2;
-          continue;
-        }
-        inQuotes = false;
-        i += 1;
-        continue;
-      }
-      field += char;
-      i += 1;
-      continue;
-    }
-
-    if (char === '"') {
-      inQuotes = true;
-      i += 1;
-      continue;
-    }
-
-    if (char === ',') {
-      pushField();
-      i += 1;
-      continue;
-    }
-
-    if (char === '\n') {
-      pushRow();
-      i += 1;
-      continue;
-    }
-
-    if (char === '\r') {
-      i += 1;
-      continue;
-    }
-
-    field += char;
-    i += 1;
-  }
-
-  if (field.length > 0 || row.length > 0) {
-    pushRow();
-  }
-
-  return rows;
-}
